greenreg: fix incorrect warning.
This fixes an incorrect warning:
When notifications are executed in an SC_METHOD GreenReg is warning about
pre-read/write notifications even if there isn't any.

So this adds a condition and warn only if there is a pre-read/write
notification.

This also fixes the wrong warning message for pre-write notification.

Signed-off-by: KONRAD Frederic <fred.konrad@greensocs.com>