greensocket: fix clang template error
include/greensocket/generic/green_socket_extension_support_base.tpp:127:14:
error: use 'template' keyword to treat 'get_extension' as a dependent template name
  T* tmp=txn.get_extension<T>(); //get the pointer
             ^
             template

ISO C++03 14.2/4

When the name of a member template specialization appears after . or -> in a
postfix-expression, or after nested-name-specifier in a qualified-id, and the
postfix-expression or qualified-id explicitly depends on a template-parameter
(14.6.2), the member template name must be prefixed by the keyword template.
Otherwise the name is assumed to name a non-template.

Signed-off-by: Guillaume Delbergue <guillaume.delbergue@greensocs.com>