gsgpsocket: fix streaming width attribute
According to the LRM: "A streaming width of 0 shall be invalid. If a streaming transfer is not required, the streaming width > attribute should be set to a value greater than or equal to the value of the data length attribute"

Set by default the streaming width to the data length value.

Signed-off-by: Guillaume Delbergue <guillaume.delbergue@greensocs.com>