greenthreads: Fix wake-up of sources that are ahead of time.
When a sync source is ahead of time (own time greater than back window
plus a quantum), we make the process sleep until the slowest source
catches up.
When SystemC was the slowest source, the wake-up was never called.
This fix makes SystemC update the backWindow and call the wake-up
function correctly.